model update

+++ endpoints/endpoints.go
@@ -57,7 +57,7 @@ type authorizeReq struct {
 		Year  int64 `json:"year"`
 		Month int64 `json:"month"`
 	} `json:"card_expiration" validate:"required,not_expired"`
-	CVV      int64  `json:"card_cvv" validate:"required,gte=100,lte=999"`
+	CVV      int32  `json:"card_cvv" validate:"required,gte=100,lte=999"`
 	Amount   int64  `json:"amount" validate:"required,gte=1"`
 	Currency string `json:"currency" validate:"required"`
 }
